@@ -1,21 +1,9 @@
 ﻿namespace Espresso.ParserDeleter.ParseRssFeeds
 {
-    public class ParseRssFeedsCommandResponse
+    public record ParseRssFeedsCommandResponse
     {
-        public int CreatedArticles { get; }
+        public int CreatedArticles { get; init; }
 
-        public int UpdatedArticles { get; }
-
-        public ParseRssFeedsCommandResponse(int createdArticles, int updatedArticles)
-        {
-            CreatedArticles = createdArticles;
-            UpdatedArticles = updatedArticles;
-        }
-
-        public override string ToString()
-        {
-            return $"{nameof(CreatedArticles)}:{CreatedArticles}, " +
-                $"{nameof(UpdatedArticles)}:{UpdatedArticles}";
-        }
+        public int UpdatedArticles { get; init; }
     }
 }
